# backend/docs/openapi.yaml
openapi: "3.0.3"
info:
  title: BidFlow API
  version: "1.0.0"
  description: |
    BidFlow is a bid/proposal management platform with ML-powered win probability prediction.
    All API routes are versioned under `/api/v1/`.

    ## Authentication
    Uses JWT tokens stored in httpOnly cookies (`access_token_cookie`, `refresh_token_cookie`).
    CSRF protection via double-submit cookie pattern (`X-CSRF-TOKEN` header).

    ## Roles
    - **Admin** — Full access to all resources
    - **Company** — Enquiry owner, can update bid status on own enquiries
    - **Sales Executive** — Internal bid management
    - **Bidder** — Submit bids on marketplace enquiries only

    ## Rate Limiting
    All endpoints are rate-limited. Limits are documented per-endpoint.
servers:
  - url: /api/v1
    description: API v1
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: access_token_cookie
      description: JWT access token in httpOnly cookie
    headerCsrf:
      type: apiKey
      in: header
      name: X-CSRF-TOKEN
      description: CSRF token from csrf_access_token cookie
  schemas:
...
    Bid:
      type: object
      properties:
        _id: { type: string }
        bidId: { type: string }
        enquiryId: { type: string }
        amount: { type: number }
        currency: { type: string, enum: [USD, EUR, GBP, INR, JPY, CAD, AUD] }
        status: { type: string, enum: [Quotation Prepared, Under Review, Negotiation, Order Received, Rejected] }
        assignedEmployee: { type: string }
        industry: { type: string }
        submissionDate: { type: string }
        remarks: { type: string }
        tags: { type: array, items: { type: string } }
        teamSize: { type: integer }
        aiPrediction: { type: integer }
        shapExplanations: { type: array, items: { $ref: '#/components/schemas/SHAPExplanation' } }
        comments: { type: array, items: { $ref: '#/components/schemas/Comment' } }
        history: { type: array, items: { $ref: '#/components/schemas/HistoryEntry' } }
        createdBy: { type: string }
...
paths:
...
  /bids/:
    get:
      tags: [Bids]
      summary: List bids
      description: Role-scoped. Admin sees all; others see assigned bids.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: page
          in: query
          schema: { type: integer, default: 1 }
        - name: size
          in: query
          schema: { type: integer, default: 50, maximum: 200 }
      responses:
        200:
          description: Paginated bid list
          content:
            application/json:
              schema:
                type: object
                properties:
                  items: { type: array, items: { $ref: '#/components/schemas/Bid' } }
                  page: { type: integer }
                  size: { type: integer }
                  total: { type: integer }

    post:
      tags: [Bids]
      summary: Create a bid
      description: Creates bid with AI prediction and SHAP explanations.
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount, enquiryId]
              properties:
                amount: { type: number, minimum: 0 }
                enquiryId: { type: string }
                currency: { type: string, enum: [USD, EUR, GBP, INR, JPY, CAD, AUD], default: USD }
                assignedEmployee: { type: string }
                industry: { type: string }
                submissionDate: { type: string }
                remarks: { type: string }
                tags: { type: array, items: { type: string } }
                teamSize: { type: integer, minimum: 1 }
                productSeries: { type: string }
                regionalOffice: { type: string }
                salesPrice: { type: number }
      responses:
        201:
          description: Bid created
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg: { type: string }
                  bid: { $ref: '#/components/schemas/Bid' }
  /bids/{id}/status:
    put:
      tags: [Bids]
      summary: Update bid status
      description: Only enquiry owner or Admin can update. Sends SHAP feedback to losing bidders. Rate limited to 30/min.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [status]
              properties:
                status: { type: string, enum: [Quotation Prepared, Under Review, Negotiation, Order Received, Rejected] }
                note: { type: string }
      responses:
        200:
          description: Status updated
        400:
          description: Invalid status
        403:
          description: Forbidden
  /bids/{id}/comments:
    post:
      tags: [Bids]
      summary: Add comment to bid
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [text]
              properties:
                text: { type: string, maxLength: 2000 }
      responses:
        201:
          description: Comment added
        400:
          description: Validation error
  /bids/{id}/comments/{commentId}:
    delete:
      tags: [Bids]
      summary: Delete a comment
      description: Only comment author, bid creator, or Admin can delete.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
        - name: commentId
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Comment deleted
        403:
          description: Forbidden
        404:
          description: Not found
  /bids/predict:
    post:
      tags: [Bids]
      summary: Get live AI prediction
      description: Returns win probability, employee win rate, and SHAP explanations.
      security: [{ cookieAuth: [] }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number }
                currency: { type: string }
                assignedEmployee: { type: string }
                industry: { type: string }
                submissionDate: { type: string }
                teamSize: { type: integer }
                productSeries: { type: string }
                regionalOffice: { type: string }
                salesPrice: { type: number }
      responses:
        200:
          description: Prediction result
          content:
            application/json:
              schema:
                type: object
                properties:
                  win_probability: { type: number }
                  computed_win_rate_pct: { type: number }
                  shap_explanations: { type: array, items: { $ref: '#/components/schemas/SHAPExplanation' } }
  /bids/calendar:
    get:
      tags: [Bids]
      summary: Get calendar events
      description: Returns bid submission dates and enquiry creation dates for calendar view.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: month
          in: query
          schema: { type: string }
          description: "Filter by month (e.g., 2026-06)"
      responses:
        200:
          description: Calendar events
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/CalendarEvent' }
  /bids/{id}:
    put:
      tags: [Bids]
      summary: Update bid fields
      description: Update tags, remarks, amount, currency, submissionDate, assignedEmployee, industry.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                tags: { type: array, items: { type: string } }
                remarks: { type: string }
                amount: { type: number }
                currency: { type: string, enum: [USD, EUR, GBP, INR, JPY, CAD, AUD] }
                submissionDate: { type: string }
                assignedEmployee: { type: string }
                industry: { type: string }
      responses:
        200:
          description: Bid updated
        400:
          description: Validation error

    delete:
      tags: [Bids]
      summary: Delete a bid
      description: Admin can always delete. Bidders can delete own bids if Rejected or 30+ days old.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Bid deleted
        403:
          description: Forbidden
  /bids/{id}/quotation:
    get:
      tags: [Bids]
      summary: Download quotation PDF
      security: [{ cookieAuth: [] }]
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: PDF file
          content:
            application/pdf:
              schema:
                type: string
                format: binary

  # ── Enquiries ────────────────────────────────────────────────────────────────
...
  /marketplace/{enquiry_id}/bid:
    post:
      tags: [Marketplace]
      summary: Submit sealed bid
      description: Only Bidder, Sales Executive, or Company roles. Deadline enforced. Rate limited to 10/min.
      security: [{ cookieAuth: [] }]
      parameters:
        - name: enquiry_id
          in: path
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [amount]
              properties:
                amount: { type: number, minimum: 0 }
                currency: { type: string, enum: [USD, EUR, GBP, INR, JPY, CAD, AUD], default: USD }
                industry: { type: string }
                submissionDate: { type: string }
                remarks: { type: string }
                tags: { type: array, items: { type: string } }
                teamSize: { type: integer, minimum: 1 }
      responses:
        201:
          description: Bid submitted
          content:
            application/json:
              schema:
                type: object
                properties:
                  msg: { type: string }
                  bid: { $ref: '#/components/schemas/Bid' }
        400:
          description: Validation error or deadline passed
        403:
          description: Forbidden

  # ── Documents ────────────────────────────────────────────────────────────────
...
  /documents/bid/{bid_id}:
    get:
      tags: [Documents]
      summary: List documents for a bid
      security: [{ cookieAuth: [] }]
      parameters:
        - name: bid_id
          in: path
          required: true
          schema: { type: string }
      responses:
        200:
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items: { $ref: '#/components/schemas/Document' }

  # ── Analytics ────────────────────────────────────────────────────────────────
...